requirements.txt
coverage==6.3.2
neat-python==0.92
numpy==1.21.5
Pillow==9.0.1
pygame==2.1.2
